SPIKE: Split cimit scoring out from gpg45profile evaluation
