+++ lib/src/main/java/uk/gov/di/ipv/core/library/domain/gpg45/Gpg45ProfileEvaluator.java
@@ -8,10 +8,6 @@ import com.nimbusds.jwt.SignedJWT;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.message.StringMapMessage;
-import uk.gov.di.ipv.core.library.domain.ContraIndicatorItem;
-import uk.gov.di.ipv.core.library.domain.ContraIndicatorMitigation;
-import uk.gov.di.ipv.core.library.domain.ContraIndicatorScore;
-import uk.gov.di.ipv.core.library.domain.ContraIndicators;
 import uk.gov.di.ipv.core.library.domain.JourneyResponse;
 import uk.gov.di.ipv.core.library.domain.gpg45.domain.CheckDetail;
 import uk.gov.di.ipv.core.library.domain.gpg45.domain.CredentialEvidenceItem;
@@ -19,8 +15,6 @@ import uk.gov.di.ipv.core.library.domain.gpg45.exception.UnknownEvidenceTypeExce
 import uk.gov.di.ipv.core.library.dto.EvidenceDto;
 import uk.gov.di.ipv.core.library.dto.Gpg45ScoresDto;
 import uk.gov.di.ipv.core.library.dto.RequiredGpg45ScoresDto;
-import uk.gov.di.ipv.core.library.exceptions.ConfigException;
-import uk.gov.di.ipv.core.library.exceptions.UnrecognisedCiException;
 import uk.gov.di.ipv.core.library.service.ConfigService;
 
 import java.text.ParseException;
@@ -29,17 +23,11 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 import java.util.stream.Collectors;
 
-import static uk.gov.di.ipv.core.library.config.ConfigurationVariable.CI_SCORING_THRESHOLD;
-import static uk.gov.di.ipv.core.library.config.CoreFeatureFlag.MITIGATION_ENABLED;
-import static uk.gov.di.ipv.core.library.domain.CriConstants.KBV_CRI;
 import static uk.gov.di.ipv.core.library.domain.VerifiableCredentialConstants.VC_CLAIM;
 import static uk.gov.di.ipv.core.library.domain.VerifiableCredentialConstants.VC_EVIDENCE;
-import static uk.gov.di.ipv.core.library.helpers.LogHelper.LogField.LOG_CI_SCORE;
 import static uk.gov.di.ipv.core.library.helpers.LogHelper.LogField.LOG_MESSAGE_DESCRIPTION;
-import static uk.gov.di.ipv.core.library.helpers.LogHelper.LogField.LOG_NO_OF_CI_ITEMS;
 import static uk.gov.di.ipv.core.library.journeyuris.JourneyUris.JOURNEY_PYI_KBV_FAIL_PATH;
 import static uk.gov.di.ipv.core.library.journeyuris.JourneyUris.JOURNEY_PYI_NO_MATCH_PATH;
 
@@ -55,90 +43,6 @@ public class Gpg45ProfileEvaluator {
 
     public Gpg45ProfileEvaluator(ConfigService configService) {
         this.configService = configService;
-    }
-
-    public Optional<JourneyResponse> getJourneyResponseForStoredContraIndicators(
-            ContraIndicators contraIndicators, boolean separateSession)
-            throws ConfigException, UnrecognisedCiException {
-        LOGGER.info(
-                new StringMapMessage()
-                        .with(LOG_MESSAGE_DESCRIPTION.getFieldName(), "Retrieved user's CI items.")
-                        .with(
-                                LOG_NO_OF_CI_ITEMS.getFieldName(),
-                                contraIndicators.getContraIndicatorsMap().size()));
-        final int ciScore =
-                contraIndicators.getContraIndicatorScore(
-                        configService.getContraIndicatorScoresMap(),
-                        configService.enabled(MITIGATION_ENABLED));
-        LOGGER.info(
-                new StringMapMessage()
-                        .with(LOG_MESSAGE_DESCRIPTION.getFieldName(), "Calculated user's CI score.")
-                        .with(LOG_CI_SCORE.getFieldName(), ciScore));
-
-        if (ciScore <= Integer.parseInt(configService.getSsmParameter(CI_SCORING_THRESHOLD))) {
-            return Optional.empty();
-        }
-
-        final String latestContraIndicatorCode =
-                contraIndicators.getLatestContraIndicator().get().getCode();
-        Map<String, ContraIndicatorMitigation> ciMitConfig = configService.getCiMitConfig();
-        if (ciMitConfig.containsKey(latestContraIndicatorCode)) {
-            final ContraIndicatorMitigation contraIndicatorMitigation =
-                    ciMitConfig.get(latestContraIndicatorCode);
-            return Optional.of(
-                    new JourneyResponse(
-                            separateSession
-                                    ? contraIndicatorMitigation.getSeparateSessionStep()
-                                    : contraIndicatorMitigation.getSameSessionStep()));
-        }
-
-        return Optional.of(JOURNEY_RESPONSE_PYI_NO_MATCH);
-    }
-
-    public Optional<JourneyResponse> getJourneyResponseForStoredCis(
-            List<ContraIndicatorItem> ciItems) throws UnrecognisedCiException {
-        List<ContraIndicatorItem> contraIndicatorItems = new ArrayList<>(ciItems);
-        LOGGER.info(
-                new StringMapMessage()
-                        .with(LOG_MESSAGE_DESCRIPTION.getFieldName(), "Retrieved user's CI items.")
-                        .with(LOG_NO_OF_CI_ITEMS.getFieldName(), ciItems.size()));
-
-        Set<String> ciSet =
-                contraIndicatorItems.stream()
-                        .map(ContraIndicatorItem::getCi)
-                        .collect(Collectors.toSet());
-
-        Map<String, ContraIndicatorScore> contraIndicatorScoresMap =
-                configService.getContraIndicatorScoresMap();
-
-        int ciScore = 0;
-        for (String ci : ciSet) {
-            if (!contraIndicatorScoresMap.containsKey(ci)) {
-                throw new UnrecognisedCiException("Unrecognised CI code received from CIMIT");
-            }
-            ContraIndicatorScore scoresConfig = contraIndicatorScoresMap.get(ci);
-            ciScore += scoresConfig.getDetectedScore();
-        }
-        LOGGER.info(
-                new StringMapMessage()
-                        .with(LOG_MESSAGE_DESCRIPTION.getFieldName(), "Calculated user's CI score.")
-                        .with(LOG_CI_SCORE.getFieldName(), ciScore));
-
-        int ciScoreThreshold =
-                Integer.parseInt(configService.getSsmParameter(CI_SCORING_THRESHOLD));
-        if (ciScore > ciScoreThreshold) {
-            Collections.sort(contraIndicatorItems);
-            String lastCiIssuer =
-                    contraIndicatorItems.get(contraIndicatorItems.size() - 1).getIss();
-            String kbvIssuer = configService.getComponentId(KBV_CRI);
-
-            return Optional.of(
-                    lastCiIssuer.equals(kbvIssuer)
-                            ? JOURNEY_RESPONSE_PYI_KBV_FAIL
-                            : JOURNEY_RESPONSE_PYI_NO_MATCH);
-        } else {
-            return Optional.empty();
-        }
     }
 
     public Optional<Gpg45Profile> getFirstMatchingProfile(
